Added Return command reporter without changing output
