+++ src/robotframework_find_unused/commands/returns/__init__.py
@@ -1,0 +1,11 @@
+"""
+Find unused keywords command
+"""
+
+from .options import ReturnOptions
+from .returns import cli_returns
+
+__all__ = [
+    "ReturnOptions",
+    "cli_returns",
+]
