+++ src/robotframework_find_unused/commands/step/keyword_count_uses.py
@@ -1,7 +1,9 @@
 from pathlib import Path
 
 from robotframework_find_unused.common.const import KeywordData, LibraryData
-from robotframework_find_unused.reporter.base.keyword_reporter import KeywordReporter
+from robotframework_find_unused.reporter.base.partial.count_keywords import (
+    PartialReporter_CountKeywords,
+)
 from robotframework_find_unused.visitors.robot import visit_robot_files
 from robotframework_find_unused.visitors.robot.keyword_visitor import RobotVisitorKeywords
 
@@ -11,7 +13,7 @@ def cli_count_keyword_uses(
     keywords: list[KeywordData],
     downloaded_libraries: list[LibraryData],
     *,
-    reporter: KeywordReporter,
+    reporter: PartialReporter_CountKeywords,
 ):
     """
     Walk through all robot files to count keyword uses and keep the user up-to-date on progress
